% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alfred_vintages}
\alias{alfred_vintages}
\title{ALFRED monthly and quarterly series vintages}
\format{
A \code{\link{list}} objects
}
\source{
\href{https://alfred.stlouisfed.org/}{ALFRED} \cr
}
\usage{
data(alfred_vintages)
}
\description{
ALFRED monthly and quarterly series vintages
}
\examples{
\donttest{
data(alfred_vintages)
i <- 1
alfred_vintages[[i]] # ith variable
}
}
\keyword{datasets}
