% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{apply_transform}
\alias{apply_transform}
\title{Time series matrix transformation}
\usage{
apply_transform(data.in, tcode.all)
}
\arguments{
\item{data.in}{time series matrix.}

\item{tcode.all}{transformation code vector. 1 - not transformed, 2 - first difference, 3 - second difference, 4 - natural log, 5 - first difference of natural log, 6 - second difference of natural log, 7 - first difference of percent change.}
}
\value{
transformed time series matrix.
}
\description{
Transform time series a matrix based on transformation code.
}
