% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_functions.R
\name{beta_w}
\alias{beta_w}
\title{Beta density polynomial weights}
\usage{
beta_w(param, dayLag)
}
\arguments{
\item{param}{two-dimensional parameter vector \eqn{\theta}.}

\item{dayLag}{number of high-frequency lags.}
}
\value{
(normalized) weights vector.
}
\description{
For a given set of parameters \eqn{\theta} and the number of high-frequency lags, returns weights implied by Beta density functional form.
}
\examples{
param <- c(2,2)
dayLag <- 22
beta_w(param, dayLag)
}
\author{
Jonas Striaukas
}
