% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cvfolds}
\alias{cvfolds}
\title{Computes cross-validation folds.}
\usage{
cvfolds(nfolds, nrow)
}
\arguments{
\item{nfolds}{(double) number of folds.}

\item{nrow}{(double) number of rows of data.}
}
\value{
foldsid (vec) vector of cv folds.
}
\description{
Computes cross-validation folds.
}
\keyword{internal}
