% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.functions.R
\name{data_freq}
\alias{data_freq}
\title{Identify data frequency}
\usage{
data_freq(DateVec)
}
\arguments{
\item{DateVec}{n by 6 vector format dates: [year,month,day,hour,min,sec]}
}
\value{
a list of arguments that define data frequency. `period` - length of two consecutive dates. `unit` - unit of length measure (=1 year;=2 month;=3 day;=4 hour;=5 minutes;=6 seconds)
}
\description{
Identify data frequency
}
\keyword{internal}
