% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{date_vec}
\alias{date_vec}
\title{Transform date vector to numeric matrix}
\usage{
date_vec(s)
}
\arguments{
\item{s}{date vector.}
}
\value{
numeric matrix of date vector.
}
\description{
Transform date vector to numeric matrix
}
\keyword{internal}
