% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.functions.R
\name{diff_time_mf}
\alias{diff_time_mf}
\title{Computes the difference between two dates.}
\usage{
diff_time_mf(
  time1,
  time2,
  origin,
  units = c("auto", "secs", "mins", "hours", "days", "weeks")
)
}
\arguments{
\item{time1}{date 1.}

\item{time2}{date 2.}

\item{origin}{date origin.}

\item{units}{units.}
}
\value{
numeric value of difference in two dates.
}
\description{
Computes the difference between two dates.
}
\keyword{internal}
