% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_functions.R
\name{expalmon_w}
\alias{expalmon_w}
\title{Exponential Almon polynomial weights}
\usage{
expalmon_w(param, dayLag)
}
\arguments{
\item{param}{two-dimensional parameter vector \eqn{\theta}.}

\item{dayLag}{number of high-frequency lags.}
}
\value{
(normalized) weights vector.
}
\description{
For a given set of parameters \eqn{\theta} and the number of high-frequency lags, returns weights implied by exponential Almon functional form.
}
\examples{
param <- c(0.02,-0.2)
dayLag <- 22
expalmon_w(param, dayLag)
}
\author{
Jonas Striaukas
}
