% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastals}
\alias{fastals}
\title{Computes fast ALS solution}
\usage{
fastals(Y, X, intercept, tau, maxIter, thresh)
}
\arguments{
\item{Y}{(vec) vector of response.}

\item{X}{(mat) matrix of covariates.}

\item{intercept}{(double) 1 - includes intercept, else not.}

\item{tau}{(double) quantile level.}

\item{maxIter}{maximum number of iterations.}

\item{thresh}{threshold value for stopping criterion.}
}
\description{
Computes fast ALS solution
}
\keyword{internal}
