% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastols}
\alias{fastols}
\title{Computes fast OLS solution}
\usage{
fastols(Y, X, intercept)
}
\arguments{
\item{Y}{(vec) vector of response.}

\item{X}{(mat) matrix of covariates.}

\item{intercept}{(double) 1 - includes intercept, else not.}
}
\description{
Computes fast OLS solution
}
\keyword{internal}
