% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastrq}
\alias{fastrq}
\title{Computes fast rq solution}
\usage{
fastrq(Y, X, intercept, tau)
}
\arguments{
\item{Y}{(vec) vector of response.}

\item{X}{(mat) matrix of covariates.}

\item{intercept}{(double) 1 - includes intercept, else not.}

\item{tau}{(double) quantile level.}
}
\description{
Computes fast rq solution
}
\keyword{internal}
