% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getmin_cpp}
\alias{getmin_cpp}
\title{Computes optimal \eqn{\lambda} for cross-validation output.}
\usage{
getmin_cpp(lambda, cvm, cvsd, which_lambda)
}
\arguments{
\item{lambda}{(vec) vector of \eqn{\lambda} parameters.}

\item{cvm}{(vec) vector of cv error curve.}

\item{cvsd}{(vec) vector of standard errors for cv error curve.}

\item{which_lambda}{which optimal \eqn{\lambda} to select; 0 - min of cv error curve, 1 - 1se min of cv error curve.}
}
\description{
Computes optimal \eqn{\lambda} for cross-validation output.
}
\keyword{internal}
