% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.sglfit.R
\name{ic.pen}
\alias{ic.pen}
\title{Compute the penalty based on chosen information criteria}
\usage{
ic.pen(ic_choice, df, t)
}
\arguments{
\item{ic_choice}{information criteria choice: BIC, AIC or AICc.}

\item{df}{effective degrees of freedom.}

\item{t}{sample size.}
}
\value{
penalty value.
}
\description{
Compute the penalty based on chosen information criteria
}
\keyword{internal}
