% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.functions.R
\name{lag_num}
\alias{lag_num}
\title{Compute the number of lags}
\usage{
lag_num(x.lag, period, unit)
}
\arguments{
\item{x.lag}{number of high-frequency lags to construct in high-frequency time units.}

\item{period}{high-frequency data period.}

\item{unit}{units.}
}
\value{
numeric value of the number of lags.
}
\description{
Compute the number of lags
}
\keyword{internal}
