% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{macro_midasml}
\alias{macro_midasml}
\title{GDP nowcasting using midasML approach example data}
\format{
A \code{\link{list}} object
}
\source{
\code{macro_midasml$rgdp} - \href{https://fred.stlouisfed.org/}{FRED} \cr
  \code{macro_midasml$md.data} - \href{https://research.stlouisfed.org/econ/mccracken/fred-databases/}{St. Louis Fed M. W. McCracken website} \cr
  \code{macro_midasml$text.data} - \href{http://www.structureofnews.com/}{The Structure of Economic News website} 
  \code{macro_midasml$survey.data} - \href{https://philadelphiafed.org/research-and-data/real-time-center/survey-of-professional-forecasters}{Philadelphia Fed SPF data}
}
\usage{
data(market_ret)
}
\description{
US real GDP, FRED-MD monthly dataset, SPF survey and textual analysis data.
}
\examples{
data(macro_midasml)
macro_midasml$rgdp.data # GDP data
macro_midasml$md.data # FRED-MD data
macro_midasml$text.data # textual analysis data
macro_midasml$survey.data # SPF data
}
\keyword{datasets}
