% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{market_ret}
\alias{market_ret}
\title{SNP500 returns}
\format{
A \code{\link{data.frame}} object.a
}
\source{
\code{rgdp} - \href{https://fred.stlouisfed.org/}{FRED}
}
\usage{
data(market_ret)
}
\description{
SNP500 returns
}
\examples{
\donttest{
data(market_ret)
market_ret$snp500ret
}
}
\keyword{datasets}
