% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.functions.R
\name{mode_midasml}
\alias{mode_midasml}
\title{Compute mode of a vector}
\usage{
mode_midasml(data)
}
\arguments{
\item{data}{data vector.}
}
\value{
a list of arguments.
}
\description{
Compute mode of a vector
}
\keyword{internal}
