% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{monthBegin}
\alias{monthBegin}
\title{Beginning of the month date}
\usage{
monthBegin(x)
}
\arguments{
\item{x}{date value.}
}
\value{
changed date value.
}
\description{
Change the date to the beginning of the month date.
}
\examples{
monthBegin(as.Date("2020-05-15"))
}
\author{
Jonas Striaukas
}
