% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{monthEnd}
\alias{monthEnd}
\title{End of the month date}
\usage{
monthEnd(x)
}
\arguments{
\item{x}{date value.}
}
\value{
changed date value.
}
\description{
Change the date to the end of the month date.
}
