% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{mse_loss}
\alias{mse_loss}
\title{Mean squared error loss for ARDL-MIDAS or DL-MIDAS regression}
\usage{
mse_loss(param, y, x, z, weight, ...)
}
\arguments{
\item{param}{parameter vector.}

\item{y}{response variable.}

\item{x}{high-frequency covariate lags.}

\item{z}{low-frequency variable lags.}

\item{weight}{non-linear weight function specification.}

\item{...}{currently ignored optional parameters.}
}
\value{
objective function value evaluated at \code{param}.
}
\description{
For a set of parameter in \code{param} vector, data inputs y, x and z, and weight function specification, computes the mse loss.
}
\keyword{internal}
