% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_functions.R
\name{plot_weights}
\alias{plot_weights}
\title{MIDAS weight plot function}
\usage{
plot_weights(obj)
}
\arguments{
\item{obj}{midas_ardl or midas_dl object with parameter estimates and model specfication inputs.}
}
\value{
returns R figure of estimated MIDAS weights.
}
\description{
Based on specification in \code{obj}, plots a basic R figure of estimated MIDAS weights.
}
\details{
MIDAS regression pecifcation is picked up from obj, see \code{midas_dl} or \code{midas_ardl} function descriptions for more details.
}
