% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sgl.R
\name{predict.panel_sgl}
\alias{predict.panel_sgl}
\title{Computes prediction for the sg-LASSO panel regression model}
\usage{
\method{predict}{panel_sgl}(object, newX, newZ = NULL, regress_choice = c("re", "fe"), ...)
}
\arguments{
\item{object}{fit object from panelsglassofit.}

\item{newX}{matrix of out-of-sample covariate observations.}

\item{newZ}{optional matrix of dummies for panel data model.}

\item{regress_choice}{choose between `re` and `fe`. Must be consistent with \code{object}.}

\item{...}{currently ignored optional parameters.}
}
\value{
a list of these variables:

pred - overall prediction.

predZ - dummies prediction.

predX - covariates prediction.
}
\description{
Computes prediction for the sg-LASSO panel regression model
}
\author{
Jonas Striaukas
}
