% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sgl.R
\name{predict.reg_sgl}
\alias{predict.reg_sgl}
\title{Computes prediction for the sg-LASSO linear regression}
\usage{
\method{predict}{reg_sgl}(object, newX, ...)
}
\arguments{
\item{object}{fit object from sglassofit.}

\item{newX}{matrix of out-of-sample covariate observations.}

\item{...}{currently ignored optional parameters.}
}
\value{
a list of these variables:

pred - overall prediction.

predZ - intercept prediction.

predX - covariates prediction.
}
\description{
Computes prediction for the sg-LASSO linear regression
}
\author{
Jonas Striaukas
}
