% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rgdp_dates}
\alias{rgdp_dates}
\title{Real GDP release dates}
\format{
A \code{\link{list}} objects
}
\source{
\href{https://alfred.stlouisfed.org/}{ALFRED} \cr
}
\usage{
data(rgdp_dates)
}
\description{
Real GDP release dates
}
\examples{
\donttest{
data(rgdp_dates)
rgdp_dates$Quarter_q # reference quarters in quarters
rgdp_dates$Quarter_m # reference quarters in months
rgdp_dates$Quarter_d # reference quarters in days
rgdp_dates$`First release` # first release date for the reference
rgdp_dates$`Second release` # second release date for the reference
rgdp_dates$`Third release` # third release date for the reference
}
}
\keyword{datasets}
