% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rgdp_vintages}
\alias{rgdp_vintages}
\title{Real GDP vintages}
\format{
A \code{\link{list}} objects
}
\source{
\href{https://alfred.stlouisfed.org/}{ALFRED} \cr
}
\usage{
data(rgdp_vintages)
}
\description{
Real GDP vintages
}
\examples{
\donttest{
data(rgdp_vintages)
rgdp_vintages$date # dates
rgdp_vintages$time_series # series, q-q annual rate
rgdp_vintages$realtime_period # real time dates 
}
}
\keyword{datasets}
