% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{transform_dt}
\alias{transform_dt}
\title{Time series vector transformation}
\usage{
transform_dt(x, tcode)
}
\arguments{
\item{x}{time series vector.}

\item{tcode}{transformation code. 1 - not transformed, 2 - first difference, 3 - second difference, 4 - natural log, 5 - first difference of natural log, 6 - second difference of natural log, 7 - first difference of percent change.}
}
\value{
transformed time series vector.
}
\description{
Transform time series a vector based on transformation code.
}
