% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscv.sglfit.R
\name{tscv.sglpath}
\alias{tscv.sglpath}
\title{Sorts time series cross-validation output}
\usage{
tscv.sglpath(outlist, lambda, x, y, foldid, ...)
}
\arguments{
\item{outlist}{list of cross-validation fits.}

\item{lambda}{a sequence of \eqn{\lambda} parameter.}

\item{x}{regressors}

\item{y}{response}

\item{foldid}{the fold assignment.}

\item{...}{other arguments passed to \code{predict.sgl}}
}
\value{
\code{cvm} and \code{cvsd}.
}
\description{
Computes \code{cvm} and \code{cvsd} based on times series cross-validation fit.
}
\keyword{internal}
