% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sgl.R
\name{updatecvmsd}
\alias{updatecvmsd}
\title{Updates the cross-validation error estimates}
\usage{
updatecvmsd(cvm, cvsd_fold, nlam, est_new, k, y_val, ZX_val)
}
\arguments{
\item{cvm}{current cvm estimates.}

\item{cvsd_fold}{current cvsd_fold estimates.}

\item{nlam}{number of \eqn{\lambda}'s to use in the regularization path.}

\item{est_new}{p by nlam matrix sequence of estimates, where p is the total dimension of the data matrix and nlam is the number of \eqn{\lambda}'s to use in the regularization path.}

\item{k}{fold index.}

\item{y_val}{response used in validation.}

\item{ZX_val}{y_val data matrix used in validation.}
}
\value{
updated cross-validation error estimates.
}
\description{
Updates the cross-validation error estimates
}
\keyword{internal}
