% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_rgdp}
\alias{us_rgdp}
\title{US real GDP data with several high-frequency predictors}
\format{
A \code{\link{list}} object.a
}
\source{
\href{https://fred.stlouisfed.org/series/GDPC1}{rgdp} \cr
  \href{https://www.chicagofed.org/research/data/cfnai/current-data}{cfnai} \cr
  \href{https://fred.stlouisfed.org/series/PAYEMS}{payems} \cr
  \href{https://www.philadelphiafed.org/research-and-data/real-time-center/business-conditions-index}{ads}
}
\usage{
data(us_rgdp)
}
\description{
US real GDP, Chicago National Activity Index, Nonfarm payrolls and ADS Index
}
\examples{
\donttest{
data(us_rgdp)
us_rgdp$rgdp # - GDP data
us_rgdp$cfnai # - CFNAI predictor data
us_rgdp$payems # - Nonfarm payrolls predictor data 
us_rgdp$ads # - ADS predictor data
}
}
\keyword{datasets}
