\name{midasml-package}
\alias{midasml-package}
\alias{midasml}
\docType{package}
\title{
\packageTitle{midasml}
}
\description{
\packageDescription{midasml}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{midasml}
\packageIndices{midasml}
midasml package contains functions to run classical MIDAS regression models under several loss functions as well as MIDAS machine learning methods. The main functions to run predictive midasml methods are reg_sgl and panel_sgl.
}
\author{
\packageAuthor{midasml}

Maintainer: \packageMaintainer{midasml}
}
\keyword{package}
