% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midasml-package.R
\docType{package}
\name{midasml-package}
\alias{midasml-package}
\alias{midasml}
\title{midasml}
\description{
Estimation and Prediction Methods for High-Dimensional Mixed Frequency Time Series Data
}
\author{
Jonas Striaukas (maintainer) \email{jonas.striaukas@gmail.com}, Andrii Babii \email{andrii@email.unc.edu}, Eric Ghysels \email{eghysels@unc.edu}
}
\keyword{package}
